pyproject.toml
# pyproject.toml
[project]
name = "gits"
version = "0.1.2"
description = "Manage multiple git repositories with group and alias support."
authors = [
  { name = "Gary Howard", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer",
    "pyyaml"
]

[project.scripts]
gits = "gits.cli:gits"
